Time: 16 ms (87.65%), Space: 22.5 MB (32.47%) - LeetHub

+++ 637-average-of-levels-in-binary-tree/637-average-of-levels-in-binary-tree.cpp
@@ -24,7 +24,7 @@ public:
                 if(curr->left) q.push(curr->left);
                 if(curr->right) q.push(curr->right);
             }
-            res.push_back((double)sum/(double)n);
+            res.push_back(sum/(double)n);
         }
         return res;
     }
